# server/openapi.yml
openapi: 3.0.3
info:
  title: ComedyInsight API
  description: |
    Full-stack mobile application backend API for comedy content insights.
    
    ## Features
    - User authentication with OTP and OAuth
    - Video content management and streaming
    - Subscription management
    - User engagement (favorites, downloads, views)
    - Admin panel for content management
    
  version: 1.0.0
  contact:
    name: ComedyInsight Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.comedyinsight.com
    description: Production server
...
paths:
...
  /downloads/request:
    post:
      tags:
        - Downloads
      summary: Request download
      description: Request download of a video and get presigned URL
      operationId: requestDownload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - video_id
                - quality
              properties:
                video_id:
                  type: string
                  format: uuid
                quality:
                  type: string
                  enum: [360p, 480p, 720p, 1080p]
                  example: 720p
      responses:
        '201':
          description: Download URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ========================================================================
  # VIEWS ENDPOINT
  # ========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from auth endpoints
...
  schemas:
...
    DownloadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        download_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
        file_size_mb:
          type: number
          format: float

    # ====================================================================
    # SUBSCRIPTION SCHEMAS
    # ====================================================================
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Validation failed'
              message:
                type: string
              details:
                type: object
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Unauthorized'
              message:
                type: string
                example: 'Authentication required'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Forbidden'
              message:
                type: string
                example: 'Insufficient permissions'
...